# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Oyna API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/differences": {
            "get": {
                "tags": [
                    "Differences"
                ],
                "summary": "Get all differences",
                "operationId": "1a602ca2958694233dce3c91bb42f9fa",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DifferenceResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Differences"
                ],
                "summary": "Create a new difference",
                "operationId": "402409e0fc7d9d8773f0f39f3d1cb839",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDifferenceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DifferenceResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/differences/{id}": {
            "get": {
                "tags": [
                    "Differences"
                ],
                "summary": "Get a specific difference",
                "operationId": "dae3633427020031dc764a8c65a16ee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DifferenceResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Differences"
                ],
                "summary": "Delete a specific difference",
                "operationId": "0d8e8e9c952b0e326a82cf0f897a7ca4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Differences"
                ],
                "summary": "Update a specific difference",
                "operationId": "daa39b97b3ec16d764d23cdb5a97d832",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDifferenceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DifferenceResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/entertainments/{entertainment}": {
            "get": {
                "tags": [
                    "Entertainments"
                ],
                "summary": "Get entertainment by id",
                "description": "Get entertainment by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "entertainment",
                        "in": "path",
                        "description": "Entertainment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entertainment retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EntertainmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entertainment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entertainment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Entertainments"
                ],
                "summary": "Delete entertainment by id",
                "description": "Delete entertainment by id",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "entertainment",
                        "in": "path",
                        "description": "Entertainment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entertainment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entertainment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entertainment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entertainment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Entertainments"
                ],
                "summary": "Update entertainment by id",
                "description": "Update entertainment by id",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "entertainment",
                        "in": "path",
                        "description": "Entertainment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "image",
                                    "city_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Entertainment name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Entertainment description"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entertainment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entertainment updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EntertainmentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entertainment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entertainment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreDifferenceRequest": {
                "required": [
                    "game_level",
                    "coordinates"
                ],
                "properties": {
                    "game_level": {
                        "description": "The level of the game",
                        "type": "string",
                        "example": "1"
                    },
                    "coordinates": {
                        "description": "The coordinates of the difference",
                        "type": "array",
                        "items": {
                            "required": [
                                "x1",
                                "y1",
                                "x2",
                                "y2"
                            ],
                            "properties": {
                                "x1": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "y1": {
                                    "type": "integer",
                                    "example": 20
                                },
                                "x2": {
                                    "type": "integer",
                                    "example": 30
                                },
                                "y2": {
                                    "type": "integer",
                                    "example": 40
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateDifferenceRequest": {
                "required": [
                    "game_level",
                    "coordinates"
                ],
                "properties": {
                    "game_level": {
                        "description": "The level of the game",
                        "type": "integer",
                        "example": 1
                    },
                    "coordinates": {
                        "description": "The coordinates of the difference",
                        "type": "array",
                        "items": {
                            "required": [
                                "x1",
                                "y1",
                                "x2",
                                "y2"
                            ],
                            "properties": {
                                "x1": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "y1": {
                                    "type": "integer",
                                    "example": 20
                                },
                                "x2": {
                                    "type": "integer",
                                    "example": 30
                                },
                                "y2": {
                                    "type": "integer",
                                    "example": 40
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "CoordinateResource": {
                "properties": {
                    "x1": {
                        "description": "The x1 coordinate",
                        "type": "integer",
                        "example": 10
                    },
                    "y1": {
                        "description": "The y1 coordinate",
                        "type": "integer",
                        "example": 20
                    },
                    "x2": {
                        "description": "The x2 coordinate",
                        "type": "integer",
                        "example": 30
                    },
                    "y2": {
                        "description": "The y2 coordinate",
                        "type": "integer",
                        "example": 40
                    }
                },
                "type": "object"
            },
            "DifferenceResource": {
                "properties": {
                    "id": {
                        "description": "The ID of the difference",
                        "type": "integer",
                        "example": 1
                    },
                    "game_level": {
                        "description": "The level of the game",
                        "type": "integer",
                        "example": 1
                    },
                    "game_id": {
                        "description": "The ID of the game",
                        "type": "integer",
                        "example": 1
                    },
                    "coordinates": {
                        "description": "The coordinates of the difference",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CoordinateResource"
                        }
                    }
                },
                "type": "object"
            },
            "EntertainmentResource": {
                "required": [
                    "id",
                    "type",
                    "title",
                    "address",
                    "description",
                    "city_id",
                    "image"
                ],
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "example": "1"
                    },
                    "type": {
                        "type": "string",
                        "example": "Entertainment"
                    },
                    "title": {
                        "type": "string",
                        "example": "Entertainment title"
                    },
                    "address": {
                        "type": "string",
                        "example": "Entertainment address"
                    },
                    "description": {
                        "type": "string",
                        "example": "Entertainment description"
                    },
                    "city_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
...
            "ImageResource": {
                "required": [
                    "id",
                    "path",
                    "imageable_type",
                    "imageable_id"
                ],
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "example": "1"
                    },
                    "path": {
                        "type": "string",
                        "example": "images/1.jpg"
                    },
                    "imageable_type": {
                        "type": "string",
                        "example": "App\\Models\\Entertainment"
                    },
                    "imageable_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}